bdd: significantly improve code sharing with MTBDD code.
From now on, the BDD and MTBDD submodules are no longer separated. Just
call either sylvan_init_bdd() or sylvan_init_mtbdd() to initialize the
BDD/MTBDD submodule.

Most of the BDD infrastructure is similar for MTBDDs. There is mostly no
need for repeated code, and the performance should be similar, except
that there is an added "leaf check" now in the garbage collection code.
I doubt that this will impact practical performance, as this touches a
cacheline that is loaded anyway.