lace: correctly use sched_getaffinity.
Checking for defined functions with defined(sched_getaffinity) is
not possible, it can only be used to check if macros are defined.

Without this commit sched_getaffinity would never be called, meaning
that lace was not able to correctly determine the number of available
cpus on systems that use e.g. docker.